+++ libs/design/src/molecules/media-gallery/media-gallery-registration.interface.ts
@@ -1,0 +1,3 @@
+export interface DaffMediaGalleryRegistration {
+	name: string;
+}
