+++ libs/design/src/molecules/media-gallery/thumbnail/thumbnail.directive.ts
@@ -11,13 +11,11 @@ import {
   OnDestroy,
 } from '@angular/core';
 
-import {
-  DaffMediaGalleryRegistration,
-  DaffThumbnailRegistration,
-  DAFF_MEDIA_GALLERY_TOKEN,
-} from '../media-gallery-token';
+import { DaffMediaGalleryRegistration } from '../media-gallery-registration.interface';
+import { DAFF_MEDIA_GALLERY_TOKEN } from '../media-gallery-token';
 import { DaffMediaGalleryRegistry } from '../registry/media-gallery.registry';
 import { daffThumbnailCompatToken } from './thumbnail-compat.token';
+import { DaffThumbnailRegistration } from './thumbnail-registration.interface';
 
 /**
  * A directive marking thumbnails for the `DaffMediaRendererComponent`. Needs to be wrapped in a `daff-media-gallery` component
